refactor(atlas): split into atlas + atlas_geometry + atlas_vpack
Splits the 4500-line monolith into three modules with clear ownership:

  nt_builder_atlas_geometry.{h,c} (~970 lines)
    Pure geometry: Point2D, convex hull, RDP, ear-clip, contour trace,
    polygon inflate (Clipper2 wrap), morphology, alpha trim, D4 transforms,
    polygon area. No builder context, no side effects beyond heap.

  nt_builder_atlas_vpack.{h,c} (~1850 lines)
    Vector packer module — formerly nt_builder_vector.inl inlined into
    atlas.c via #include hack. Now a real .c with its own header.
    Owns the shared types AtlasPlacement, PackStats, ATLAS_MAX_PAGES
    (vpack is the sole producer, atlas.c reads them for BENCH logging).
    Also absorbs AreaSortEntry/area_sort_cmp and atlas_trace_str_ieq/
    env_truthy which are vpack-only consumers post-tile_pack.

  nt_builder_atlas.c (~1700 lines, was 4500)
    Pipeline orchestrator, dedup, geometry stage, compose, debug PNG,
    cache read/write, serialize, register, public begin/add/end API,
    pack_stats_measure_payload (the only stats helper that lives in
    pipeline land, computes fill ratio after vpack returns).

Tests: TestPoint2D typedef removed; tests now include
nt_builder_atlas_geometry.h and call public geometry functions
directly. The 8 nt_atlas_test_* wrapper layer is gone — only
nt_atlas_test_vpack_point_in_nfp survives because vpack internals
remain static (it stays in atlas_vpack.c).

Internal cleanup along the way:
  - vpack_cross and vpack_rect_overlap deleted (truly unused, were
    masked by .inl-inlining; tidy now sees them in their own TU).
  - vector_pack definition was 'static' which contradicted the new
    public header — un-staticed.
  - Duplicate Point2D typedef in atlas_geometry.c removed.
  - nt_builder_internal.h test wrapper declarations removed.
  - File header comment in atlas.c updated for new layout.

Build clean. All 28 tests pass. clang-format clean. clang-tidy delta
-4 errors (vpack_cross + vpack_rect_overlap + vector_pack-static +
duplicate Point2D), no new errors introduced.
